split out underscore methods
